+++ src/app/layout/toolbar/toolBar.styl
@@ -4,8 +4,8 @@
     display flex
     align-items center
 
-//.toolbar-tools
-//  display flex
+.toolbar-tools
+  display flex
   &__item
     margin 0 8px !important
     &:first-child
