+++ src/shared/components/Notification/notification.icons.ts
@@ -1,0 +1,14 @@
+import { NotificationType } from 'shared/components/Notification/notification.types'
+import {
+  CheckCircle as CheckCircleIcon,
+  Error as ErrorIcon,
+  Info as InfoIcon,
+  Warning as WarningIcon,
+} from '@material-ui/icons'
+
+export const icons: { [type: NotificationType]: JSXElement } = {
+  [NotificationType.success]: CheckCircleIcon,
+  [NotificationType.warning]: WarningIcon,
+  [NotificationType.error]: ErrorIcon,
+  [NotificationType.info]: InfoIcon,
+}
