+++ src/shared/components/Notification/notification.types.ts
@@ -1,0 +1,6 @@
+export enum NotificationType {
+  warning = 'warning',
+  error = 'error',
+  info = 'info',
+  success = 'success',
+}
